Two fixes $1, $2
$1: changed incorrect usage of sizeof(test_path) in memset(). Sizeof()
was incorrectly used to return the size of the allocated array, but is
incapable of doing so for dynamically allocated memory. The parameter
for the call to malloc() is used instead.

$2: For all malloc() calls the sizeof() calls were switched to use a
dereference of the pointer itself rather than the type of the pointer.
Eg. sizeof(*ptr) is used instead of sizeof(char) for allocating memory.
This is to add resilence to the code if the type of the pointer is
changed.